requirements.txt
aiofiles
chardet
gradio_client
langdetect
mattermostdriver-asyncio
openai
pylint
tiktoken
webuiapi
